+++ src/events/PdfEvent.php
@@ -14,9 +14,17 @@ use yii\base\Event;
  * Class PdfEvent
  *
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
- * @since 4.0
+ * @since 2.0
  */
 class PdfEvent extends Event
 {
+    /**
+     * @var Pdf The PDF model associated with the event.
+     */
     public Pdf $pdf;
+
+    /**
+     * @var bool Whether the PDF is brand new
+     */
+    public bool $isNew = false;
 }