+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes for Craft Commerce
 
+## Unreleased
+
+### Fixed
+- Fixed a PHP that occurred when making a payment on a gateway with no form parameters.
+
 ## 4.0.0-beta.2 - 2022-04-06
 
 ### Changed