+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## Unreleased
 
+### Added
+- Added `craft\commerce\events\PdfRenderEvent`.
+- Added `craft\commerce\services\Pdfs::EVENT_BEFORE_DELETE_PDF`.
+
+## Changed
+- `craft\commerce\services\Pdfs::EVENT_BEFORE_RENDER_PDF` now raises a `craft\commerce\events\PdfRenderEvent` rather than a `\craft\commerce\events\PdfEvent`.
+- `craft\commerce\services\Pdfs::EVENT_AFTER_RENDER_PDF` now raises a `craft\commerce\events\PdfRenderEvent` rather than a `\craft\commerce\events\PdfEvent`.
+- `craft\commerce\services\Pdfs::EVENT_BEFORE_SAVE_PDF` now raises a `craft\commerce\events\PdfEvent` rather than a `\craft\commerce\events\PdfSaveEvent`.
+- `craft\commerce\services\Pdfs::EVENT_AFTER_SAVE_PDF` now raises a `craft\commerce\events\PdfEvent` rather than a `\craft\commerce\events\PdfSaveEvent`.
+
+### Removed
+- Removed `\craft\commerce\events\PdfSaveEvent`.
+ 
 ### Fixed
 - Fixed addresses not showing as primary on the Edit Address slideout. ([#2762](https://github.com/craftcms/commerce/issues/2762))
 - Fixed a bug that could occur causing field layout tabs not to show on the Edit Order page. ([#2770](https://github.com/craftcms/commerce/issues/2770))